requirements.txt
streamlit
qrcode
pillow
pandas
numpy
python-dateutil
